Pathfinding
The 'loop' folder is required for the pathfinding to work as it contains the priority queue structure. 
Programmed an A* algorithm in lua, also added a small test file with an example use.
Tested it on my own computer and didn't encounter any issues. 

If you find an error or want to suggest a change, let me know so I can fix it.