MultiLML (#57)
Added a function for creating multiplications and a function for checking fairness of created graph.

Multiplications are created by iterating through all possible connections, therefore the creation time goes to extremely long for 8 or more players (or a lot of connections in each node).

It is worth noting that it is possible that there doesn't exist a fair multiplication for given node, a simple example:
MultiplyLML({{1, 'LOCAL', 2}, {2, 'LOCAL, 3}, {3 'LOCAL', 4}}, 3), you can easily see that it is impossible to create a 'fair' match for 3 players with such a node. 

I also only assumed what it means that a connection is 'fair', therefore feel free to tell me if I got it wrong and give me an example where current rules wont find an 'unfair' connection.